+++ push-docker-images.sh
@@ -21,7 +21,7 @@ done
 
 VERSION=$(grep -m1 '__version__' app.py | sed 's/.*"\(.*\)".*/\1/')
 
-MODELS=(pii_base hospitals medical_conditions french_persons french_medical)
+MODELS=(pii_en_small pii_en_xsmall pii_en_medium pii_en_large pii_base hospitals medical_conditions french_persons french_medical)
 
 for MODEL in "${MODELS[@]}"; do
     TAG="philterd/ph-eye:${VERSION}-${MODEL}"